fix(图表): 修复子弹图横轴标签长度限制失效的问题
